add 'count' as valid aggregation method
